requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
tortoise-orm[asyncpg]>=0.22.0
pydantic>=2.0.0
pyyaml>=6.0
msgpack>=1.0
pycryptodome>=3.20.0
httpx>=0.27.0
aiofiles>=24.0
